Split Action Button to Multiple Component
